pkg/repo: read remote KEP files from clone
The remote KEP loading code was using the local repo as its base and
failing to find files. This updates the code to take two parameters, the
path to the repo, and the path to the KEP in the repo.

This allows `kepctl query --include-prs` to work (again? I'm not sure if
it used to work and was broken during refactors)

While I was here, I renamed ReadKEP to LoadLocalKEP (for parity with
LoadLocalKEPs)